+++ src/options/components/General/ContextMenus.tsx
@@ -4,6 +4,7 @@ import { observer } from 'mobx-react';
 import { translator } from '../../../common/translator';
 import { rootStore } from '../../stores';
 import { ControlsSwitch } from '../ui/Controls';
+import { Prefs } from '../../../common/prefs';
 
 export const ContextMenus = observer(() => {
     const { settingsStore } = useContext(rootStore);
@@ -12,6 +13,14 @@ export const ContextMenus = observer(() => {
     const handleToggle = async (): Promise<void> => {
         await settingsStore.setContextMenusValue(!contextMenusEnabled);
     };
+
+    /**
+     * Oculus Browser does not support context menus yet (AG-44721).
+     * Remove this check when it will be available.
+     */
+    if (Prefs.isOculus()) {
+        return null;
+    }
 
     return (
         <ControlsSwitch